+++ test/scheduler.js
@@ -8,18 +8,17 @@ test('enque && tick', t => {
   let n = 0
   s.enque(5, () => t.equals(++n, 2, 'Second'))
   s.enque(3, () => t.equals(++n, 1, 'First'))
-  s.enque(15, () => t.equals(++n, 4, 'Fourth'))
+  s.enque(15, () => { t.equals(++n, 4, 'Fourth'); t.end() })
   s.enque(7, () => t.equals(++n, 3, 'Third'))
 
   t.equals(s.toArray().map(i => i.at).join(','), '3,5,7,15')
   let i = 0
   while (!s.empty) s.tick(i++)
-  t.end()
 })
 
 test('cancel', t => {
   t.plan(4)
-  const s = new Scheduler()
+  const s = new Scheduler(false)
   let n = 0
   s.enque(2, () => t.equals(++n, 1, 'First'))
   const cancel = s.enque(4, t.fail.bind(null, 'Should have been canceled'))
@@ -30,5 +29,4 @@ test('cancel', t => {
   t.equals(s.toArray().map(i => i.at).join(','), '2,5,13')
   let i = 0
   while (!s.empty) s.tick(i++)
-  t.end()
 })